+++ ai-post-scheduler/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [wizard-authors-search] - 2025-12-27
+### Added
+- Added client-side search functionality to the Authors list admin page.
+
 ## [wizard-clone-template] - 2025-01-01
 ### Added
 - Added "Clone Template" functionality to allow users to easily duplicate templates with a single click.